Sprint2 Task49 Separate integration and unit tests
-add maven failsafe plugin to pom for skipping integration test when mvn
package
-add builder helper plugin to pom to possible of creating new package test-integration
-mock tests which suppose to be a unit tests: BookDataDownloaderTest
-add tests for BookService class
