+++ src/test/java/edition/academy/seventh/serivce/BookDataDownloaderTest.java
@@ -5,20 +5,23 @@ import org.testng.annotations.Test;
 
 import java.util.List;
 
-/**
- * @author Ola Podorska
- */
+import static org.mockito.Mockito.*;
+
+/** @author Ola Podorska */
 @Test
 public class BookDataDownloaderTest {
 
   @Test(
-      dataProviderClass = DataProviderForJsonParser.class,
+      dataProviderClass = DataProviderForConvertingIsbnToBook.class,
       dataProvider = "dataProviderForConvertingIsbnToBook")
   public void should_returnListOfBooksAsString_when_haveListOfIsbns(
       List<String> isbns, List<String> expectedBooks) {
     // Given
+    BookstoreConnector bookstoreConnector = mock(BookstoreConnector.class);
+    when(bookstoreConnector.getJsonResponse(anyString())).thenReturn(expectedBooks.get(0));
+
     BookDataDownloader bookDataDownloader =
-        new BookDataDownloader("https://api.itbook.store/1.0/books/", new BookstoreConnector());
+        new BookDataDownloader("https://api.itbook.store/1.0/books/", bookstoreConnector);
 
     // When
     List<String> books = bookDataDownloader.listBooksByIsbn(isbns);
