+++ src/test/java/edition/academy/seventh/serivce/BookMapperTest.java
@@ -14,7 +14,7 @@ import static org.testng.Assert.assertEquals;
  * @author Bartosz Kupajski
  * @author Kamil Rojek
  */
-public class ITBookMapperTest {
+public class BookMapperTest {
   private static List<Book> bookList;
 
   @BeforeTest
@@ -44,7 +44,8 @@ public class ITBookMapperTest {
   }
 
   @Test(dataProvider = "dataProviderForJSONMapping")
-  public void should_returnActualBookProperties_when_fromValidJson(String bookPart, String expected) {
+  public void should_returnActualBookProperties_when_fromValidJson(
+      String bookPart, String expected) {
     assertEquals(bookPart, expected);
   }
 }
